Extract out stock chart into its own component so it can be used in other places
