+++ frontend/src/app/stock-chart.component/stock-chart.component.css
@@ -1,0 +1,9 @@
+canvas {
+  width: 100% !important;
+  height: 100% !important;
+}
+
+.chart-wrapper {
+  height: 250px;
+  width: 100%;
+}
