+++ frontend/src/app/stock-chart.component/stock-chart.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { StockChartComponent } from './stock-chart.component';
+
+describe('StockChartComponent', () => {
+  let component: StockChartComponent;
+  let fixture: ComponentFixture<StockChartComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [StockChartComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(StockChartComponent);
+    component = fixture.componentInstance;
+    await fixture.whenStable();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
