+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "bootstrap": "^5.3.8"
+  }
+}
